+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "Go-Web Framework" will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.8.0-beta] - 2021-09-02
+### Changed
+- Abstract CLI command structure
+- Improved kernel registry removing useless declaration.
+- Implemented HTTP router in base entities
 
 ## [v0.7.1-beta] - 2021-09-02
 ### Fixed